refactor: simplify switch networks
